Split WMRNationalPolicies
